Update README.md
In this game, the player can move the player (black box) left or right using the left or right arrow keys. The goal is to prevent the ball (blue square) from falling to the floor, which would cause the ball to reverse direction when it hits the floor. The ball will automatically move down and back up when it hits the floor.